+++ cartagen/algorithms/points/__init__.py
@@ -1,2 +1,2 @@
-from cartagen.algorithms.points.quadtree import *
-from cartagen.algorithms.points.reduction import *
+from cartagen.algorithms.points.covering import hull_delaunay, hull_swinging_arm
+from cartagen.algorithms.points.reduction import reduce_kmeans, reduce_quadtree, reduce_labelgrid