+++ cartagen/algorithms/points/reduction.py
@@ -2,11 +2,15 @@
 # A reduction means that there is still a set of points as output of the algorithm, but fewer points.
 
 from shapely.geometry import MultiPoint, Point, Polygon
-from cartagen.algorithms.points.quadtree import PointSetQuadTree
+from cartagen.utils.partitioning.quadtree import PointSetQuadTree
 import random
+import pandas as pd
+import geopandas as gpd
+import numpy as np
+import matplotlib.pyplot as plt
 
 # @gtouya
-def reduce_points_kmeans(points, shrink_ratio, centroid=False):
+def reduce_kmeans(points, shrink_ratio, centroid=False):
     """
     Reduce a set of points using K-Means clustering.
 
@@ -30,8 +34,10 @@ def reduce_points_kmeans(points, shrink_ratio, centroid=False):
 
     See Also
     --------
-    reduce_points_quadtree :
+    reduce_quadtree :
         Reduce a set of points using a quadtree.
+    reduce_labelgrid :
+        Reduce a set of points using the Label Grid method.
     """
 
     final_pts = []
@@ -105,7 +111,7 @@ def reduce_points_kmeans(points, shrink_ratio, centroid=False):
 
     return final_pts
 
-def reduce_points_quadtree(points, depth, mode='simplification', attribute=None):
+def reduce_quadtree(points, depth, mode='simplification', attribute=None):
     """
     Reduce a set of points using a quadtree.
 
@@ -147,8 +153,10 @@ def reduce_points_quadtree(points, depth, mode='simplification', attribute=None)
 
     See Also
     --------
-    reduce_points_kmeans :
+    reduce_kmeans :
         Reduce a set of points using K-Means clustering.
+    reduce_labelgrid :
+        Reduce a set of points using the Label Grid method.
 
     References
     ----------
@@ -232,4 +240,223 @@ def reduce_points_quadtree(points, depth, mode='simplification', attribute=None)
     
     return output, qtree
 
+def reduce_labelgrid(points, attribute, width, height, shape='square', mode='selection', grid=False):
+    """
+    Reduce a set of points using the Label Grid method.
 
+    This algorithm was proposed by Gröbe & Burghardt. :footcite:p:`grobe:2021`
+    It assigns each point to a grid cell of a given height, width and shape
+    before reducing the points per cells either by selection or aggregation.
+    
+    Parameters
+    ----------
+    points : GeoDataFrame of Point
+        The points to reduce.
+    attribute : str
+        Name of the attribute to rank the points.
+    width : float
+        Width of the grid cells.
+    height : float
+        Height of the grid cells.
+    shape : str, optional
+        Shape of the grid cells, can be 'square', 'diamond', 'hexagonal'.
+    mode : str, optional
+        The reduction method used, can be:
+
+        - **'selection'** keeps the point with the highest attribute value.
+        - **'aggregation'** keeps the centroid of the cell.
+    grid : bool, optional
+        If set to True, returns a tuple with the points and the grid.
+
+    Returns
+    -------
+    GeoDataFrame of Point or tuple (GeoDataFrame of Point, GeoDataFrame of Polygon)
+
+    See Also
+    --------
+    reduce_kmeans :
+        Reduce a set of points using K-Means clustering.
+    reduce_quadtree :
+        Reduce a set of points using a quadtree.
+
+    References
+    ----------
+    .. footbibliography::
+    """
+    lg = LabelGrid(points, attribute, width, height, shape, mode)
+    lg.set_point_label_grid()
+    result = lg.getPointResults()
+    if grid:
+        rgrid = lg.getGrid()
+        return result, rgrid
+    else:
+        return result
+
+"""
+Created on Sun Jun 16 22:40:42 2024.
+
+@author: vpech
+"""
+class LabelGrid():
+    """
+    A class for label grid algorithm.
+    
+    The algorithm was proposed by Gröbe, M. and Burghardt, D., 2021. 
+    Scale-Dependent Point Selection Methods for Web Maps. KN - Journal of 
+    Cartography and Geographic Information.
+    """
+    
+    def __init__(self, points, attribute, width, height, typ='square', mode='selection'):
+        """
+        Construct all the necessary attributes for the LabelGrid object.
+
+        Parameters
+        ----------
+        points : GeoDataFrame
+            GeoDataFrame of object points we want to apply the label grid algorithm.
+        width : int
+            Width of a cell.
+        height : int
+            Height of a cell.
+        typ : str
+            Form of a cell. The default is 'carre'.
+        """
+        self.__attribute = attribute
+        self.__width = width
+        self.__height = height
+        self.__typ = typ
+        self.__mode = mode
+        self.__points = points
+        self.__points_res = None
+        self.__grid = None
+
+    def __create_hexagonalCell(self, coords_min, coords_max):
+        """
+        Create a grid of hexagonal cells.
+
+        Parameters
+        ----------
+        coords_min : tuple
+            Coordinates of the bottom left cell first point.
+        coords_max : tuple
+            Coordinates of the top right cell last point.
+
+        Returns
+        -------
+        polygons : GeoDataFrame
+            Contains hexagonal polygon objects.
+        """
+        xstep = self.__width*3
+        ystep = self.__width*2
+        cols = list(np.arange(coords_min[0], coords_max[0]+xstep, xstep))
+        rows = list(np.arange(coords_min[1], coords_max[1]+ystep, ystep))
+        
+        polygons = [Polygon([(x, y), 
+                             (x + self.__width, y),
+                             (x+self.__width*3/2, y+self.__width),
+                             (x + self.__width, y+2*self.__width),
+                             (x, y+2*self.__width),
+                             (x-self.__width/2, y+self.__width)]) 
+                    for x in cols[:-1] for y in rows[:-1]]
+        return polygons
+
+    def __createGrid(self):
+        """
+        Create the grid of the label grid.
+
+        Returns
+        -------
+        GeoDataFrame
+            Contains the grid polygons.
+
+        """
+        xmin, ymin, xmax, ymax = self.__points.total_bounds
+        
+        if self.__typ == 'square':
+            cols = list(np.arange(xmin, xmax + self.__width, self.__width))
+            rows = list(np.arange(ymin, ymax + self.__height, self.__height))
+        
+            polygons = [Polygon([(x, y), 
+                                 (x + self.__width, y), 
+                                 (x + self.__width, y + self.__height), 
+                                 (x, y + self.__height)]) 
+                        for x in cols[:-1] for y in rows[:-1]]
+            
+        if self.__typ == 'diamond':
+            cols = list(np.arange(xmin, xmax + self.__width*2, self.__width))
+            rows = list(np.arange(ymin-self.__width, ymax + self.__width, self.__width))
+            
+            polygons = [Polygon([(x, y), 
+                                 (x + self.__width / 2, y + self.__width / 2), 
+                                 (x, y + self.__width), 
+                                 (x - self.__width / 2, y + self.__width / 2)]) 
+                        for x in cols[:-1] for y in rows[:-1]]
+            
+        if self.__typ == 'hexagonal':
+
+            odd_coords_min = (xmin-self.__width, ymin-self.__width)
+            odd_coords_max = (xmax+self.__width, ymax)
+            
+            odd_poly = self.__create_hexagonalCell(odd_coords_min, odd_coords_max)
+            
+            not_odd_coords_min = (odd_coords_min[0]+3/2*self.__width, odd_coords_min[1]-self.__width)
+            not_odd_coords_max = (odd_coords_max[0], odd_coords_max[1]+self.__width)
+            
+            not_odd_poly = self.__create_hexagonalCell(not_odd_coords_min, not_odd_coords_max)
+            
+            polygons = odd_poly + not_odd_poly
+            
+        return gpd.GeoDataFrame({'geometry':polygons})
+
+    def set_point_label_grid(self):
+        """Set the attributes points_res and grid of the LabelGrid object."""
+        self.__points.reset_index(drop=True, inplace=True)
+        self.__points = self.__points[['geometry', self.__attribute]]
+        
+        self.__grid = self.__createGrid()
+        
+        lst_in_value = [self.__points.loc[cell.contains(self.__points['geometry']), self.__attribute] 
+                        for cell in self.__grid['geometry']]
+
+        lst_inter_value = [self.__points.loc[cell.touches(self.__points['geometry']), self.__attribute] 
+                           for cell in self.__grid['geometry']]
+        
+        lst_all_value = [lst_inter_value[i] 
+                         if lst_in_value[i].empty 
+                         else pd.concat([lst_in_value[i],lst_inter_value[i]]) 
+                         for i in range(len(lst_in_value))]
+        
+        if self.__mode == 'selection':
+            ind = [e.nlargest(1).index[0] for e in lst_all_value if not e.empty]
+            p = [self.__points['geometry'].iloc[i] for i in ind]
+            point_results = gpd.GeoDataFrame(geometry=gpd.GeoSeries(p))
+            
+        if self.__mode == 'aggregation':
+            aggreg = [(geom, len(cell)) for geom, cell in zip(self.__grid.centroid, lst_all_value) if not cell.empty]
+            df = pd.DataFrame(aggreg, columns=["geometry", "radius"])
+            point_results = gpd.GeoDataFrame(df, geometry="geometry")
+            
+        self.__points_res = point_results
+
+    def getPointResults(self):
+        """Get points results."""
+        return self.__points_res
+    
+    def getGrid(self):
+        """Get the grid."""
+        return self.__grid
+    
+    def draw(self):
+        """Draw the grid with the points given and the grid with point results."""
+        fig, ax1 = plt.subplots()
+        self.__grid.boundary.plot(ax=ax1, color='goldenrod')
+        self.__points.plot(ax=ax1, color='black')
+        
+        fig, ax2 = plt.subplots()
+        self.__grid.boundary.plot(ax=ax2, color='goldenrod')
+        
+        if self.__mode == 'selection':
+            self.__points_res.plot(ax=ax2, color='chocolate')
+            
+        if self.__mode == 'aggregation':
+            self.__points_res.plot(ax=ax2, color='chocolate', markersize=self.__width*self.__points_res["radius"])