+++ docs/conf.py
@@ -7,9 +7,9 @@
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
 import os, sys
-import cartagen
-
 sys.path.insert(0, os.path.abspath('..'))
+
+import cartagen
 
 project = 'cartagen'
 copyright = '2024, IGN, Univ Gustave Eiffel'